requirements.txt
streamlit>=1.28.0
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python-headless>=4.8.0
PyYAML>=6.0
pandas>=2.0.0
rapidfuzz>=3.0.0
numpy>=1.24.0
